Fix issue where PR without any description has its title edited
This came up in https://github.com/python/cpython/pull/5291

The PR did not have any description, and the title was
edited to include the bpo number.
It caused a 500 error, because bedevere did not anticipate
this situation.

Updated the code so that it takes into consideration when the
body is empty.