CI: set `fail-fast` to False
With `fail-fast` set to True (the default), all running jobs in the matrix are cancelled as soon as a single one fails. This makes it hard to see exactly which Python versions a PR is failing on, if there are failures, because the CI just becomes a sea of red, even if the PR is only failing on a single Python version.