gh-556: Fix over-reaching PR-edited event handler
gh-555 introduced the PR edited handler to handle draft PRs.
The implementation was over-reaching, and triggered on any edited events (including renaming PR).
This PR fixes that by removing the PR-edited handler and replacing it with
two more specific handlers that more narrowly match the original intent,
namely, respond to changes in the draft state of the PR (publish / unpublish).

Fixes gh-556